+++ packages/php-wasm/node-builds/8-5/asyncify/php_8_5.js
@@ -31329,6 +31329,27 @@ export function init(RuntimeName, PHPLoader) {
 		});
 	};
 
+	function ___emscripten_lookup_name(namePtr) {
+		if (!ENVIRONMENT_IS_NODE) {
+			return original__emscripten_lookup_name(namePtr);
+		}
+		if (!PHPLoader.syscalls) {
+			return original__emscripten_lookup_name(namePtr);
+		}
+
+		const hostname = UTF8ToString(namePtr);
+
+		let ipString = '';
+		try {
+			ipString = PHPLoader.syscalls.gethostbyname(hostname);
+		} catch (e) {
+			// Fall through to the default synthetic mapping if native DNS fails.
+		}
+
+		return inetPton4(ipString);
+	}
+	___emscripten_lookup_name.sig = 'ip';
+
 	var webSockets = new HandleAllocator();
 
 	var WS = {
@@ -31748,6 +31769,10 @@ export function init(RuntimeName, PHPLoader) {
 	// invocation, so that we will immediately be able to queue the newest
 	// produced audio samples.
 	registerPostMainLoop(() => SDL.audio?.queueNewAudioData?.());
+	const original__emscripten_lookup_name = __emscripten_lookup_name;
+	if (typeof __emscripten_lookup_name !== 'undefined') {
+		__emscripten_lookup_name = ___emscripten_lookup_name;
+	}
 	// End JS library code
 
 	// include: postlibrary.js
@@ -31783,6 +31808,9 @@ export function init(RuntimeName, PHPLoader) {
 	Module['addRunDependency'] = addRunDependency;
 	Module['removeRunDependency'] = removeRunDependency;
 	Module['ccall'] = ccall;
+	Module['UTF8ToString'] = UTF8ToString;
+	Module['stringToUTF8'] = stringToUTF8;
+	Module['lengthBytesUTF8'] = lengthBytesUTF8;
 	Module['FS_preloadFile'] = FS_preloadFile;
 	Module['FS_unlink'] = FS_unlink;
 	Module['FS_createPath'] = FS_createPath;
@@ -32762,6 +32790,8 @@ export function init(RuntimeName, PHPLoader) {
 		__cxa_rethrow: ___cxa_rethrow,
 		/** @export */
 		__cxa_throw: ___cxa_throw,
+		/** @export */
+		__emscripten_lookup_name: ___emscripten_lookup_name,
 		/** @export */
 		__resumeException: ___resumeException,
 		/** @export */